requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pyusb==1.2.1
pybluez==0.23
bluetooth-python==0.22
psutil==5.9.5
netifaces==0.11.0
python-nmap==0.7.1
pystray==0.19.4
pillow==10.0.0
plyer==2.1.0
pyserial==3.5
websockets==11.0.3
zeroconf==0.112.0
cryptography==41.0.4
schedule==1.2.0
watchdog==3.0.0
configparser==6.0.0